+++ src/main/java/hust/tools/hmm/io/BinaryFileHMMWriter.java
@@ -35,7 +35,7 @@ public class BinaryFileHMMWriter extends AbstractHMMWriter {
 
 	
 	@Override
-	public void writeCount(long count) throws IOException {
+	public void writeCount(int count) throws IOException {
 		dos.writeLong(count);
 	}
